Update factory and move to subtypes
